add controllers and repositories
