--- Unifyx_Backend/src/main/java/org/example/unifyx/Entities/WorkerProfile.java
@@ -1,62 +1,0 @@
-package org.example.unifyx.Entities;
-
-import jakarta.persistence.*;
-
-import java.util.List;
-
-@Entity
-public class WorkerProfile {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int workerId;
-
-    private String name;
-    private String email;
-    private String phoneNo;
-
-    @OneToMany(mappedBy = "worker", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<Portfolio> portfolios;
-
-    // Getters and Setters
-
-    public int getWorkerId() {
-        return workerId;
-    }
-
-    public void setWorkerId(int workerId) {
-        this.workerId = workerId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public List<Portfolio> getPortfolios() {
-        return portfolios;
-    }
-
-    public void setPortfolios(List<Portfolio> portfolios) {
-        this.portfolios = portfolios;
-    }
-
-    public String getPhoneNo() {
-        return phoneNo;
-    }
-
-    public void setPhoneNo(String phoneNo) {
-        this.phoneNo = phoneNo;
-    }
-}
-
