+++ Unifyx_Backend/src/main/java/org/example/unifyx/Model/OwnerProfile.java
@@ -1,0 +1,77 @@
+package org.example.unifyx.Model;
+
+import jakarta.persistence.*;
+import java.util.List;
+
+@Entity
+public class OwnerProfile {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int ownerId;
+
+    private String name;
+    private String email;
+    private String phoneNo;
+    private String address;
+    private String siteAddress;
+
+    @OneToMany(mappedBy = "owner", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Post> posts;
+
+    // Getters and Setters
+    public int getOwnerId() {
+        return ownerId;
+    }
+
+    public void setOwnerId(int ownerId) {
+        this.ownerId = ownerId;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhoneNo() {
+        return phoneNo;
+    }
+
+    public void setPhoneNo(String phoneNo) {
+        this.phoneNo = phoneNo;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public String getSiteAddress() {
+        return siteAddress;
+    }
+
+    public void setSiteAddress(String siteAddress) {
+        this.siteAddress = siteAddress;
+    }
+
+    public List<Post> getPosts() {
+        return posts;
+    }
+
+    public void setPosts(List<Post> posts) {
+        this.posts = posts;
+    }
+}
