+++ Unifyx_Backend/src/main/java/org/example/unifyx/Repository/ContractorProfileRepository.java
@@ -1,0 +1,10 @@
+package org.example.unifyx.Repository;
+
+import org.example.unifyx.Model.ContractorProfile;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ContractorProfileRepository extends JpaRepository<ContractorProfile, Integer> {
+    // You can add custom queries here if necessary
+}
