+++ frontend/app/build.gradle
@@ -34,6 +34,7 @@ dependencies {
     implementation 'com.google.firebase:firebase-auth'
     implementation 'com.google.firebase:firebase-firestore'
     implementation 'com.google.firebase:firebase-storage'
+    
 
     implementation libs.appcompat
     implementation libs.material
