+++ frontend/app/src/main/java/com/example/unifyx/network/RetrofitClient.java
@@ -1,0 +1,19 @@
+package com.example.unifyx.network;
+
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class RetrofitClient {
+    private static Retrofit retrofit;
+    private static final String BASE_URL = "http://10.0.2.2:8080/";  // Replace with your actual backend IP
+
+    public static Retrofit getRetrofitInstance() {
+        if (retrofit == null) {
+            retrofit = new Retrofit.Builder()
+                    .baseUrl(BASE_URL)
+                    .addConverterFactory(GsonConverterFactory.create())
+                    .build();
+        }
+        return retrofit;
+    }
+}
